Fixups

+++ litex/soc/software/libbase/qsort.c
@@ -196,15 +196,15 @@ loop:
   pn = (char *) base + nmemb * size;
   r = min(pa - (char *)base, pb - pa);
   vecswap(base, pb - r, r);
-  r = min(pd - pc, pn - pd - size);
+  r = min(pd - pc, pn - pd - (int)size);
   vecswap(pb, pn - r, r);
 
-  if ((r = pb - pa) > size)
+  if ((r = pb - pa) > (int)size)
     {
       qsort(base, r / size, size, compar);
     }
 
-  if ((r = pd - pc) > size)
+  if ((r = pd - pc) > (int)size)
     {
       /* Iterate rather than recurse to save stack space */
       base = pn - r;
